semantic-image-search-backend/requirements.txt
fastapi==0.110.2
uvicorn[standard]==0.29.0
torch==2.3.0
torchvision==0.18.0
pillow==10.3.0
transformers==4.41.2
accelerate==0.30.1
numpy==1.26.4
scikit-learn==1.4.2
python-dotenv==1.0.1
pinecone-client==3.2.2
pydantic-settings==2.3.0
httpx==0.27.0
pytest==8.2.1
